backend/requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
boto3==1.28.6
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
httpx==0.27.2
pyOpenSSL==25.0.0
